# docs/swagger.yaml
basePath: /
definitions:
  models.AddUserToTenantRequest:
    properties:
      role_id:
        type: string
      tenant_id:
        type: string
      user_id:
        type: string
    required:
    - role_id
    - tenant_id
    - user_id
    type: object
  models.CreateItemMeta:
    properties:
      message:
        example: Item created successfully
        type: string
      table:
        example: customers
        type: string
    type: object
  models.CreateItemResponse:
    properties:
      data:
        additionalProperties: true
        type: object
      meta:
        $ref: '#/definitions/models.CreateItemMeta'
    type: object
  models.CreateTenantRequest:
    properties:
      domain:
        type: string
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
  models.DeleteItemMeta:
    properties:
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      message:
        example: Item deleted successfully
        type: string
      table:
        example: customers
        type: string
    type: object
  models.DeleteItemResponse:
    properties:
      meta:
        $ref: '#/definitions/models.DeleteItemMeta'
    type: object
  models.DeleteTenantResponse:
    properties:
      message:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        example: INVALID_TABLE
        type: string
      details:
        example: Table 'invalid_table' does not exist or is not accessible
        type: string
      error:
        example: Invalid table name
        type: string
    type: object
  models.ItemMeta:
    properties:
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      table:
        example: customers
        type: string
    type: object
  models.ItemResponse:
    properties:
      data:
        additionalProperties: true
        type: object
      meta:
        $ref: '#/definitions/models.ItemMeta'
    type: object
  models.ItemsListMeta:
    properties:
      count:
        example: 25
        type: integer
      limit:
        example: 25
        type: integer
      offset:
        example: 0
        type: integer
      table:
        example: customers
        type: string
      total:
        example: 100
        type: integer
      type:
        example: data
        type: string
    type: object
  models.ItemsListResponse:
    properties:
      data:
        items:
          additionalProperties: true
          type: object
        type: array
      meta:
        $ref: '#/definitions/models.ItemsListMeta'
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      tenant_slug:
        description: 'Optional: specify tenant for login'
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      tenant_id:
        type: string
      tenant_slug:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.SignUpRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      tenant_slug:
        description: 'Optional: create/join tenant during signup'
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  models.SignUpResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.SwitchTenantRequest:
    properties:
      tenant_id:
        type: string
    required:
    - tenant_id
    type: object
  models.SwitchTenantResponse:
    properties:
      message:
        type: string
      tenant_id:
        type: string
      tenant_slug:
        type: string
      token:
        type: string
    type: object
  models.Tenant:
    properties:
      created_at:
        type: string
      domain:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  models.TenantResponse:
    properties:
      message:
        type: string
      tenant:
        $ref: '#/definitions/models.Tenant'
    type: object
  models.UpdateItemMeta:
    properties:
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      table:
        example: customers
        type: string
    type: object
  models.UpdateItemResponse:
    properties:
      data:
        additionalProperties: true
        type: object
      meta:
        $ref: '#/definitions/models.UpdateItemMeta'
    type: object
  models.UpdateTenantRequest:
    properties:
      domain:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      slug:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
      updated_at:
        type: string
    type: object
  models.UserTenantResponse:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
  description: Directus-style API with Role-Based Access Control (RBAC). A powerful,
    generic API that provides CRUD operations for any database table with comprehensive
    security, multi-tenancy, and dynamic schema management.
  title: Basin API
  version: 1.0.0
paths:
  /auth/context:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Auth Context
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/me:
    get:
      description: Retrieve information about the currently authenticated user. Requires
        valid JWT Bearer token or API key.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sign up payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SignUpResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sign Up
      tags:
      - auth
  /auth/switch-tenant:
    post:
      consumes:
      - application/json
      parameters:
      - description: Switch tenant payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.SwitchTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SwitchTenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Switch Tenant
      tags:
      - auth
  /auth/tenants:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tenant'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get User Tenants
      tags:
      - auth
...
  /items/{table}:
    get:
      description: Retrieve a paginated list of items from any dynamic table in the
        system. This endpoint works with both core schema tables (users, roles, permissions,
        collections, fields, api-keys) and custom dynamic tables (e.g., blog_posts,
        customers, products). The API automatically adapts to the table's schema,
        applying filters, sorting, and pagination. Requires authentication via JWT
        Bearer token or API key.
      parameters:
      - description: Table name (e.g., 'users', 'blog_posts', 'customers')
        in: path
        name: table
        required: true
        type: string
      - description: Limit (max 500, default 25)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Page number (1-based, alternative to offset)
        in: query
        name: page
        type: integer
      - description: Items per page (alternative to limit)
        in: query
        name: per_page
        type: integer
      - description: Sort field (e.g., 'created_at', 'name', 'email')
        in: query
        name: sort
        type: string
      - description: 'Sort order: ASC or DESC (default: DESC)'
        in: query
        name: order
        type: string
      - description: JSON filter object for advanced filtering
        in: query
        name: filter
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Page (1-based)
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: ASC or DESC
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ItemsListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List items from dynamic table
      tags:
      - items
    post:
      consumes:
      - application/json
      description: Create a new item in any dynamic table in the system. This endpoint
        works with both core schema tables and custom dynamic tables. The item structure
        depends on the table's schema (fields, validation rules, etc.). Requires authentication
        via JWT Bearer token or API key.
      parameters:
      - description: Table name (e.g., 'users', 'blog_posts', 'customers')
        in: path
        name: table
        required: true
        type: string
      - description: Item data
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create item in dynamic table
      tags:
      - items
  /items/{table}/{id}:
    delete:
      description: Delete an item from any dynamic table in the system. This endpoint
        works with both core schema tables and custom dynamic tables. The deletion
        is permanent and cannot be undone. Requires authentication via JWT Bearer
        token or API key.
      parameters:
      - description: Table name (e.g., 'users', 'blog_posts', 'customers')
        in: path
        name: table
        required: true
        type: string
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Delete item from dynamic table
      tags:
      - items
    get:
      description: Retrieve a specific item by ID from any dynamic table in the system.
        This endpoint works with both core schema tables and custom dynamic tables.
        Requires authentication via JWT Bearer token or API key.
      parameters:
      - description: Table name (e.g., 'users', 'blog_posts', 'customers')
        in: path
        name: table
        required: true
        type: string
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Get item from dynamic table
      tags:
      - items
    put:
      consumes:
      - application/json
      description: Update an existing item in any dynamic table in the system. This
        endpoint works with both core schema tables and custom dynamic tables. Only
        the fields provided in the request body will be updated. Requires authentication
        via JWT Bearer token or API key.
      parameters:
      - description: Table name (e.g., 'users', 'blog_posts', 'customers')
        in: path
        name: table
        required: true
        type: string
      - description: Item ID
        in: path
        name: id
        required: true
        type: string
      - description: Item data to update
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpdateItemResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Update item in dynamic table
      tags:
      - items
...
  /items/users:
    get:
      description: Retrieve a list of users in the system. Requires authentication
        and user management permissions.
      parameters:
      - description: Limit (max 500)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Page (1-based)
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      - description: Sort field (email, first_name, last_name, created_at)
        in: query
        name: sort
        type: string
      - description: ASC or DESC
        in: query
        name: order
        type: string
      - description: Filter by email
        in: query
        name: email
        type: string
      - description: Filter by active status
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ItemsListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user in the system. Requires authentication and user
        creation permissions.
      parameters:
      - description: User data (email, first_name, last_name, password, role_id, tenant_id)
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Create user
      tags:
      - users
...
  /tenants:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tenant'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get All Tenants
      tags:
      - tenants
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant creation payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Tenant
      tags:
      - tenants
  /tenants/{id}:
    delete:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteTenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete Tenant
      tags:
      - tenants
    get:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tenant'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Tenant by ID
      tags:
      - tenants
    put:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tenant'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update Tenant
      tags:
      - tenants
  /tenants/{id}/join:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserTenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Join Tenant (Current User)
      tags:
      - tenants
  /tenants/{id}/users:
    post:
      consumes:
      - application/json
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Add user payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.AddUserToTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserTenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add User to Tenant
      tags:
      - tenants
  /tenants/{id}/users/{user_id}:
    delete:
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserTenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove User from Tenant
      tags:
      - tenants
securityDefinitions:
  ApiKeyAuth:
    description: 'API key for programmatic access (format: Bearer YOUR_API_KEY)'
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    description: JWT Bearer token for user authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"